api/requirements.txt
Django==6.0.3
djangorestframework==3.15.2
django-cors-headers==4.8.0
google-generativeai==0.8.3
python-dotenv==1.0.1
